Sales migration - replace bcid data type from int to bigInteger

+++ database/migrations/2023_08_10_095240_create_sales_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('so_no', 25);
             $table->string('si_no', 25)->nullable();
             $table->string('si_assignment_id', 25)->nullable();
-            $table->integer('bcid');
+            $table->bigInteger('bcid');
             $table->string('distributor_name');
             $table->double('shipping_fee')->default(0);
             $table->double('total_amount')->default(0);
